avcodec/mlp_parser: split off shared code to its own file
